requirements.txt
langchain_community
langchain_openai
langchain_text_splitters
langchain_classic
langchain_core
azure-search-documents
azure-identity
langchain
openai
six
operator
fastapi
uvicorn